vector-embedding/requirements.txt
uvicorn==0.21.1
fastapi==0.95.1
openai==1.11.1
matplotlib==3.7.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
spacy==3.6.0
PyMuPDF==1.23.2
easyocr==1.7.0
SpeechRecognition==3.10.0
plotly==5.14.1
pydub==0.25.1
python-multipart==0.0.6
pandas
bs4==0.0.1
networkx==2.8.8
rdflib==7.0.0
pypdf
langchain
tiktoken
schedule==1.2.1
transformers
accelerate==0.22.0
einops==0.6.1
xformers==0.0.21
bitsandbytes==0.41.3
sentence_transformers==2.2.2
faiss-gpu
google-cloud-aiplatform